+++ example/example_app.dart
@@ -8,11 +8,17 @@ class ExampleApp extends StatelessWidget {
    Widget build(BuildContext context) {
      return Scaffold(
        body: Center(
-         child: CircleCard(
-           radius: 50.0,
-           elevation: 8.0,
-           backgroundColor: Colors.amber ,
-           child: Image.asset("assets/image.jpg"),
+         child: Column(
+           crossAxisAlignment: CrossAxisAlignment.center,
+           mainAxisAlignment: MainAxisAlignment.center,
+           children: [
+             CircleCard(
+               radius: 100.0,
+               elevation: 8.0,
+               backgroundColor: Colors.amber ,
+               child: Image.asset("assets/image.jpg"),
+             ),
+           ],
          ),
        ),
      );